Add support for id* type parameters
1、Add support for id* type parameters. See the `funcTestNSErrorPointer`
example in `test.js`.
If you are calling a method which have id* parameters, you can pass a
pointer to it.
But there are two things should be remembered.
Remember to memset the pointer to zero before using it, and use `pval`
to get the object pointed by the pointer.
Remember to call `releaseTmpObj` manually  if you don’t need the object
any more.

2、 Modified `- (id)formatOCToJS:`.  Unified the return value of C API
Extensions and Objective-C method.